+++ lib/hooks/queries/usePortfolioPositions.ts
@@ -36,7 +36,10 @@ import { calcResolvedMarketPrices } from "lib/util/calc-resolved-market-prices";
 import { useMemo } from "react";
 import { MarketBond, useAccountBonds } from "./useAccountBonds";
 import { useChainTime } from "lib/state/chaintime";
-import { ForeignAssetPrices, useAllAssetUsdPrices } from "./useAssetUsdPrice";
+import {
+  ForeignAssetPrices,
+  useAllForeignAssetUsdPrices,
+} from "./useAssetUsdPrice";
 
 export type UsePortfolioPositions = {
   /**
@@ -147,12 +150,11 @@ export const usePortfolioPositions = (
 ): UsePortfolioPositions => {
   const now = useChainTime();
 
-  //todo: needs to base asset balance?
   const { data: ztgPrice } = useZtgPrice();
   const block24HoursAgo = Math.floor(now?.block - 7200);
   const { data: marketBonds, isLoading: isBondsLoading } =
     useAccountBonds(address);
-  const { data: foreignAssetPrices } = useAllAssetUsdPrices();
+  const { data: foreignAssetPrices } = useAllForeignAssetUsdPrices();
 
   const rawPositions = useAccountTokenPositions({
     where: {
@@ -586,10 +588,6 @@ export const usePortfolioPositions = (
 
 /**
  * Calculates the total value of a set of positions in ZTG
- *
- * @param positions positions: Position[]
- * @param key "price" | "price24HoursAgo"
- * @returns Decimal
  */
 export const totalPositionsValue = <
   K extends keyof Pick<Position, "price" | "price24HoursAgo">,
